Map is moved to the new place
